tst_timer: Fix compilation on pre C11 compilers
The anonymous unions are supported since C11 and we have to still
support at least gcc-4.4 which does not support C11 since the compiler
is older than the standard.

So this commit gives the union name however 99% of code is unchanged
since we are using function to access the structure members anyways.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Reported-by: Petr Vorel <pvorel@suse.cz>