fanotify: Move safe_fanotify_init() + safe macro into fanotify.h
fanotify code is used only in testcases/kernel/syscalls/fanotify/,
which justify breaking rule of having safe function and macro outside
of the library.

The original motivation was to do further modifications: have libc
wrapper and raw syscall() in .test_variants for fanotify_init() and
fanotify_mark(). As both have been included in glibc for some time we
drop this idea, but it still make sense to have all fanotify helpers
in a single place.

Acked-by: Amir Goldstein <amir73il@gmail.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Li Wang <liwang@redhat.com>
Signed-off-by: Petr Vorel <pvorel@suse.cz>