syscalls/mmap: add test for MAP_FIXED_NOREPLACE
This patch adds a new test case for the mmap syscall. It tests the
MAP_FIXED_NOREPLACE flag of mmap. The code checks if MAP_FIXED_NOREPLACE
returns with EEXIST when mapped with an already mapped address. It does
so by allocating an available address by passing NULL to first argument
of mmap and tries to mmap with MAP_FIXED_NOREPLACE flag at the same
address returned by the first mmap call. This fails as expected. It also
does the necessary changes required to run the syscall using the runltp
command after building the LTP test suite.

Resolves #299
Signed-off-by: Kushal Chand <kushalchand@zilogic.com>
Reviewed-by: Vijay Kumar B. <vijaykumar@zilogic.com>
Acked-by: Jan Stancek <jstancek@redhat.com>