testcases.mk: Fix LTPLIBS build rule
The current code that attempts to rebuild libraries a test is depending
on is broken for more than one library since the "$^" is a list of all
prerequisities.

This fixes the problem by adding a proper pattern rule.

I've also added the libraries to the .PHONY target so that we attempt to
rebuild them regardless of the library archive existence since we do not
know the prerequisities of the library archive in the testcase directory
we cannot tell if some of the sources has been changed so the safest
course of action is to attempt the rebuild every time.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>