syscalls/futex_wait03: Fix synchronization
When I was reviewing changes in the test I noticed that the
synchronization was never working as the getppid() returns the test
library pid and not the thread that we should wait for.

So this fixes the test by passing the correct pid casted to a void* as
a parameter to the threaded function.

It seems that in practice this cannot be triggered easily, but the
test without this fix fails easily when you insert bussy loop
such as "for (volatile int i = 0; i < 1000000000; i++);" between
the SAFE_PTHRED() and the futex_wait() call.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>