Update mmap06 and mmap07 to work with sgx-lkl's mmap implementation
As part of changing sgx-lkl to only use it's internal mmap implementation,
we lost the ability to distinguish between EBADF and EACCES errors that
mmap06, mmap07, and mmap08 were testing the distinction between.

Overall, the corresponding sgx-lkl change is a win as it removes a source
of subtle bugs. A corresponding issue for sgx-lkl is open to track this
discrepancy in the mmap API so that in the future, these changes could
be reverted.