lib/tst_clocks.c: fix 'tst_clock_gettime()' and 'tst_clock_settime()'
The type of 'struct tst_ts' has to be kept between function calls
because a function can be called with the wrong timespec type.
For that reason, a 'static' identifier is added for 'tts'
in 'tst_clock_gettime()' and 'tst_clock_settime()' as to make sure
the right timespec type is passed in 'clock_gettime()' and
'clock_settime()'.

Signed-off-by: Filip Bozuta <Filip.Bozuta@syrmia.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>