Add a test case for mmap MAP_GROWSDOWN flag
This test implements two cases.

First one uses MAP_GROWSDOWN mapping as a thread stack and expects that
the thread will grow the stack successfully.

The second one will have a 'break' page allocated into the space the
mapping is supposed to grow and expects that the thread will be killed
with SIGSEGV.

Resolves #300
Signed-off-by: Pravin Raghul S. <pravinraghul@zilogic.com>
Reviewed-by: Vijay Kumar B. <vijaykumar@zilogic.com>
Signed-off-by: Li Wang <liwang@redhat.com>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Li Wang <liwang@redhat.com>