IMA: Rewrite ima_boot_aggregate.c to new API
The main reason was to see TCONF messages, which are printed into stderr
in new API (but to stdout in legacy API) and thus visible as the output
is redirected into the variable.

Changing boot_aggregate: to sha1: to be compatible with evmctl
ima_boot_aggregate.

Also require root (needed for reading
/sys/kernel/security/tpm0/binary_bios_measurements).

Reviewed-by: Mimi Zohar <zohar@linux.ibm.com>
Signed-off-by: Petr Vorel <pvorel@suse.cz>