madvise06: Add tag for "mm: memcg: link page counters to root ..."
The test may trigger a warning if bf4f059 is present (5.9) and
use_hierarchy=0 which is fixed by 8de15e920dc8. Full details of the
bug are in the kernel commit description and I do not want to repeat
them in the test description as they are likely to be irrelevant to
most readers.

Signed-off-by: Richard Palethorpe <rpalethorpe@suse.com>
Reviewed-by: Li Wang <liwang@redhat.com>