lib/tst_kconfig: Make use of boolean expression eval
Now each string in the kconfig[] array in tst_test structure is an
boolean expression which is evaluated. All expressions has to be true in
order for the test to continue.

This also makes the parser for the kernel config a bit more robust as it
was pointed out that there may have been cases where it could be mislead
by hand edited config files.

+ Update the docs.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
CC: Pengfei Xu <pengfei.xu@intel.com>
Reviewed-by: Xiao Yang <yangx.jy@cn.fujitsu.com>