fanotify: Add helper for FAN_REPORT_FID support on fs
and use it in relevant tests setup().
The purpose is to reduce checks (cleanup).

FAN_REPORT_FID can be unsupported in kernel entirely (detected in
fanotify_init()) or just missing for particular fs (detected in
fanotify_mark().

The check is related to kernel fix
a8b13aa20afb ("fanotify: enable FAN_REPORT_FID init flag")

NOTE: FAN_REPORT_FID is missing (at least) on exfat and ntfs over FUSE.
Original motivation for cleanup in this and previous commits was not
only to simplify the code but also check for unsupported filesystems
which was missing least for exfat on fanotify16, because filesystem can
be set via LTP_DEV_FS_TYPE environment variable on tests which don't
tests all filesystems.

Suggested-by: Amir Goldstein <amir73il@gmail.com>
Reviewed-by: Amir Goldstein <amir73il@gmail.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>
Signed-off-by: Petr Vorel <pvorel@suse.cz>